Remove catalog.Interface as we don't put it in any use
