+++ pkg/samplerepo/samplecontent.go
@@ -25,7 +25,7 @@ const (
 	sampleRepoCommitMsg  = "Add sample data"
 )
 
-func PopulateSampleRepo(ctx context.Context, repo *catalog.Repository, cat catalog.Interface, pathProvider upload.PathProvider, blockAdapter block.Adapter, user *model.User) error {
+func PopulateSampleRepo(ctx context.Context, repo *catalog.Repository, cat *catalog.Catalog, pathProvider upload.PathProvider, blockAdapter block.Adapter, user *model.User) error {
 	// upload sample data
 	// we skip checking if the repo and branch exist, since we just created them
 	// we also skip checking if the file exists, since we know the repo is empty
@@ -117,7 +117,7 @@ func PopulateSampleRepo(ctx context.Context, repo *catalog.Repository, cat catal
 	return err
 }
 
-func AddBranchProtection(ctx context.Context, repo *catalog.Repository, cat catalog.Interface) error {
+func AddBranchProtection(ctx context.Context, repo *catalog.Repository, cat *catalog.Catalog) error {
 	return cat.SetBranchProtectionRules(ctx, repo.Name, &graveler.BranchProtectionRules{
 		BranchPatternToBlockedActions: map[string]*graveler.BranchProtectionBlockedActions{
 			repo.DefaultBranch: {Value: []graveler.BranchProtectionBlockedAction{graveler.BranchProtectionBlockedAction_COMMIT}},
